Moved dragdrop to utils folder
